requirements.txt
django-crispy-forms==1.9.1
et-xmlfile==1.0.1
jdcal==1.4.1
numpy==1.18.5
openpyxl==2.6.4
pandas==0.24.2
psycopg2==2.8.5
psycopg2-binary==2.8.5
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
XlsxWriter==1.2.9
